+++ changelog.md
@@ -1,3 +1,8 @@
-## Jan 30, 2021
+# Changelog of PicoScenes Plugin Development Kit (PDK)
+
+## Feb. 19, 2021
+    * [EchoProbe] Stop Tx/Rx service when initiator / injector job finishes.
+
+## Jan. 30, 2021
     * add changelog
     * improve round-trip measurement responsiveness for SDR